+++ projects/bjoern-plugins/vsa/src/main/java/bjoern/plugins/vsa/structures/StridedInterval64Bit.java
@@ -7,7 +7,7 @@ final class StridedInterval64Bit extends StridedInterval
 	static final StridedInterval64Bit TOP = new StridedInterval64Bit(1, DataWidth.R64.getMinimumValue(),
 			DataWidth.R64.getMaximumValue());
 
-	StridedInterval64Bit(int stride, long lower, long upper)
+	StridedInterval64Bit(long stride, long lower, long upper)
 	{
 		super(stride, lower, upper);
 	}
